+++ examples/tests/async_runner.rs
@@ -5,10 +5,8 @@ use alloc::collections::{BTreeMap, BTreeSet};
 use manul::{
     dev::{BinaryFormat, TestSessionParams, TestSigner},
     protocol::Protocol,
-    session::{
-        signature::Keypair, CanFinalize, LocalError, Message, RoundOutcome, Session, SessionId, SessionParameters,
-        SessionReport,
-    },
+    session::{CanFinalize, LocalError, Message, RoundOutcome, Session, SessionId, SessionParameters, SessionReport},
+    signature::Keypair,
 };
 use manul_example::simple::{SimpleProtocol, SimpleProtocolEntryPoint};
 use rand::Rng;