+++ CHANGELOG.md
@@ -3,6 +3,18 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.2.0] - in development
+
+### Changed
+
+- Removed `protocol::digest` re-export ([#75]).
+- `digest` and `signature` are now re-exported from the top level instead of `session` ([#75]).
+
+
+[#75]: https://github.com/entropyxyz/manul/pull/75
+
+
 ## [0.1.0] - 2024-11-19
 
 ### Changed